Consolidates compat modules
