--- dask_ml/model_selection/_compat.py
@@ -1,6 +1,0 @@
-from distutils.version import LooseVersion
-from sklearn import __version__
-
-_SK_VERSION = LooseVersion(__version__)
-
-_HAS_MULTIPLE_METRICS = _SK_VERSION >= '0.19.0'
